Fix config modules being imported in the event loop
There was a late import in this integration because of the circular import.

The code has been rearranged to avoid the circular imports
